adserver/requirements.txt
aioredis==1.3.1
asgiref==3.2.7
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
daphne==2.5.0
Django==3.0.5
django-cors-headers==3.2.1
django-filter==2.3.0
django-rest-swagger==2.2.0
django-summernote==0.8.11.6
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
inflection==0.4.0
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
openapi-codec==1.3.2
Pillow==7.1.2
psycopg2==2.8.5
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
pytz==2019.3
requests==2.23.0
service-identity==18.1.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.11.2
zope.interface==5.1.0
